fix: more Clippy and no-std compilation fixes
Clippy errors in --all-targets:
- tests/bridge_fusion.rs: remove unused PipeError, remove usize-to-usize cast
- tests/lossless_jpeg.rs: use struct update syntax instead of
  field-assignment-after-default (collapsible_default_struct_construction)

no_std test compilation errors:
- tests/coverage_fill.rs: gate analyze tests with #[cfg(feature = "std")]
  (AnalyzeBuilder takes 2 args with std, 1 arg without)
- src/graph.rs: #[allow(dead_code)] on no-std UpstreamMeta struct

Example requires std:
- Cargo.toml: add [[example]] required-features = ["std"] for bench_windowed
  (uses WindowedFilterSource and zenfilters which are std-only)

+++ tests/bridge_fusion.rs
@@ -18,14 +18,14 @@ use zennode::*;
 use zenpipe::bridge::{self, NodeConverter};
 use zenpipe::graph::NodeOp;
 use zenpipe::sources::MaterializedSource;
-use zenpipe::{PipeError, Source, format};
+use zenpipe::{Source, format};
 
 // ═══════════════════════════════════════════════════════════════════════
 // Test infrastructure
 // ═══════════════════════════════════════════════════════════════════════
 
 fn solid_source(w: u32, h: u32) -> Box<dyn Source> {
-    let bpp = format::RGBA8_SRGB.bytes_per_pixel() as usize;
+    let bpp = format::RGBA8_SRGB.bytes_per_pixel();
     let data = vec![128u8; w as usize * h as usize * bpp];
     Box::new(MaterializedSource::from_data(
         data,
